technical_test/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
packaging==23.2
psycopg2==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
